Update TON White Paper link in README
The commit updates the TON White Paper link in the README file to a new URL.
Update TON White Paper link in README

The commit updates the TON White Paper link in the README file to a new URL. The previous URL was replaced with a new one that points to a different location where the white paper can be accessed. This change ensures that users are directed to the correct resource when they click on the link.
Update TON White Paper link in README

The commit updates the TON White Paper link in the README file to a new URL. The previous URL was replaced with a new one that points to a different location where the white paper can be accessed. This change ensures that users are directed to the correct resource when they click on the link.
Update TON White Paper link in README

The commit updates the TON White Paper link in the README file to a new URL. The previous URL was replaced with a new one that points to a different location where the white paper can be accessed. This change ensures that users are directed to the correct resource when they click on the link.